@@ -1612,6 +1612,7 @@ abstract class HelixMinOffsetSetByDomainsAll
       _$helixMinOffsetSetByDomainsAllSerializer;
 }
 
+// when we want each helix to have its own max based on its domains
 abstract class HelixMaxOffsetSetByDomainsAll
     with BuiltJsonSerializable, UndoableAction
     implements Action, Built<HelixMaxOffsetSetByDomainsAll, HelixMaxOffsetSetByDomainsAllBuilder> {
@@ -1622,6 +1623,21 @@ abstract class HelixMaxOffsetSetByDomainsAll
 
   static Serializer<HelixMaxOffsetSetByDomainsAll> get serializer =>
       _$helixMaxOffsetSetByDomainsAllSerializer;
+}
+
+// when we want the same max to be applied to all helices
+abstract class HelixMaxOffsetSetByDomainsAllSameMax
+    with BuiltJsonSerializable, UndoableAction
+    implements
+        Action,
+        Built<HelixMaxOffsetSetByDomainsAllSameMax, HelixMaxOffsetSetByDomainsAllSameMaxBuilder> {
+  /************************ begin BuiltValue boilerplate ************************/
+  factory HelixMaxOffsetSetByDomainsAllSameMax() = _$HelixMaxOffsetSetByDomainsAllSameMax;
+
+  HelixMaxOffsetSetByDomainsAllSameMax._();
+
+  static Serializer<HelixMaxOffsetSetByDomainsAllSameMax> get serializer =>
+      _$helixMaxOffsetSetByDomainsAllSameMaxSerializer;
 }
 
 abstract class HelixOffsetChangeAll
@@ -1791,13 +1807,17 @@ abstract class ConvertCrossoverToLoopout
 
   int get length;
 
+  @nullable
+  String get dna_sequence;
+
   StrandPart get strand_part => crossover;
 
   /************************ begin BuiltValue boilerplate ************************/
-  factory ConvertCrossoverToLoopout(Crossover crossover, int length) =>
+  factory ConvertCrossoverToLoopout(Crossover crossover, int length, [String dna_sequence = null]) =>
       ConvertCrossoverToLoopout.from((b) => b
         ..crossover.replace(crossover)
-        ..length = length);
+        ..length = length
+        ..dna_sequence = dna_sequence);
 
   factory ConvertCrossoverToLoopout.from([void Function(ConvertCrossoverToLoopoutBuilder) updates]) =
       _$ConvertCrossoverToLoopout;
@@ -1869,6 +1889,29 @@ abstract class JoinStrandsByCrossover
   JoinStrandsByCrossover._();
 
   static Serializer<JoinStrandsByCrossover> get serializer => _$joinStrandsByCrossoverSerializer;
+}
+
+// used to move a linker (crossover or loopout, stored as potential_crossover.linker)
+// so that one end stays fixed (stored in potential_crossover.dna_end_first_clicked)
+// while the other end moves to dna_end_second_click, editing two strands
+abstract class MoveLinker
+    with BuiltJsonSerializable, UndoableAction
+    implements Action, Built<MoveLinker, MoveLinkerBuilder> {
+  PotentialCrossover get potential_crossover;
+
+  DNAEnd get dna_end_second_click;
+
+  /************************ begin BuiltValue boilerplate ************************/
+  @memoized
+  int get hashCode;
+
+  factory MoveLinker({PotentialCrossover potential_crossover, DNAEnd dna_end_second_click}) = _$MoveLinker._;
+
+  factory MoveLinker.from([void Function(MoveLinkerBuilder) updates]) = _$MoveLinker;
+
+  MoveLinker._();
+
+  static Serializer<MoveLinker> get serializer => _$moveLinkerSerializer;
 }
 
 // JoinStrandsByCrossover cannot be in a BatchAction since the reducer for it looks up strands
@@ -2187,8 +2230,12 @@ abstract class DomainsMoveStartSelectedDomains
     implements Action, Built<DomainsMoveStartSelectedDomains, DomainsMoveStartSelectedDomainsBuilder> {
   Address get address;
 
+  BuiltMap<int, int> get original_helices_view_order_inverse;
+
   /************************ begin BuiltValue boilerplate ************************/
-  factory DomainsMoveStartSelectedDomains({Address address}) = _$DomainsMoveStartSelectedDomains._;
+  factory DomainsMoveStartSelectedDomains(
+      {Address address,
+      BuiltMap<int, int> original_helices_view_order_inverse}) = _$DomainsMoveStartSelectedDomains._;
 
   DomainsMoveStartSelectedDomains._();
 
@@ -2811,6 +2858,31 @@ abstract class ModificationRemove
   static Serializer<ModificationRemove> get serializer => _$modificationRemoveSerializer;
 }
 
+abstract class ModificationConnectorLengthSet
+    with BuiltJsonSerializable
+    implements Action, Built<ModificationConnectorLengthSet, ModificationConnectorLengthSetBuilder> {
+  Strand get strand;
+
+  Modification get modification;
+
+  int get connector_length;
+
+  /************************ begin BuiltValue boilerplate ************************/
+  @memoized
+  int get hashCode;
+
+  factory ModificationConnectorLengthSet({Strand strand, Modification modification, int connector_length}) =
+      _$ModificationConnectorLengthSet._;
+
+  factory ModificationConnectorLengthSet.from(
+      [void Function(ModificationConnectorLengthSetBuilder) updates]) = _$ModificationConnectorLengthSet;
+
+  ModificationConnectorLengthSet._();
+
+  static Serializer<ModificationConnectorLengthSet> get serializer =>
+      _$modificationConnectorLengthSetSerializer;
+}
+
 /////////////////////////////////////////////////////////////////////////////////////////////////////////////
 // modification edit
 
@@ -3251,7 +3323,7 @@ abstract class AutofitSet with BuiltJsonSerializable implements Action, Built<Au
 }
 
 /////////////////////////////////////////////////////////////////////////////////////////////////////////////
-// show or hide helix circles/text in main view
+// show or hide helix circles/text, or entire helix (including offset lines) in main view
 
 abstract class ShowHelixCirclesMainViewSet
     with BuiltJsonSerializable
@@ -3264,6 +3336,26 @@ abstract class ShowHelixCirclesMainViewSet
   ShowHelixCirclesMainViewSet._();
 
   static Serializer<ShowHelixCirclesMainViewSet> get serializer => _$showHelixCirclesMainViewSetSerializer;
+}
+
+abstract class ShowHelixComponentsMainViewSet
+    with BuiltJsonSerializable
+    implements Action, Built<ShowHelixComponentsMainViewSet, ShowHelixComponentsMainViewSetBuilder> {
+  bool get show_helix_components;
+
+  /************************ begin BuiltValue boilerplate ************************/
+  @memoized
+  int get hashCode;
+
+  factory ShowHelixComponentsMainViewSet({bool show_helix_components}) = _$ShowHelixComponentsMainViewSet._;
+
+  factory ShowHelixComponentsMainViewSet.from(
+      [void Function(ShowHelixComponentsMainViewSetBuilder) updates]) = _$ShowHelixComponentsMainViewSet;
+
+  ShowHelixComponentsMainViewSet._();
+
+  static Serializer<ShowHelixComponentsMainViewSet> get serializer =>
+      _$showHelixComponentsMainViewSetSerializer;
 }
 
 /////////////////////////////////////////////////////////////////////////////////////////////////////////////
