@@ -1,11 +1,24 @@
+import 'dart:html';
 import 'dart:math';
 
+import 'package:meta/meta.dart';
 import 'package:over_react/over_react.dart';
+import 'package:built_collection/built_collection.dart';
 
+import '../state/strand.dart';
+import '../state/address.dart';
 import '../state/geometry.dart';
-import '../state/domain.dart';
-import 'pure_component.dart';
+import '../state/group.dart';
+import '../app.dart';
 import '../state/helix.dart';
+import '../state/domain.dart';
+import '../util.dart' as util;
+import '../state/selectable.dart';
+import 'design_main_strand_dna_end.dart';
+import 'pure_component.dart';
+import '../state/context_menu.dart';
+import '../state/modification_type.dart';
+import '../actions/actions.dart' as actions;
 import '../constants.dart' as constants;
 
 part 'design_main_strand_domain_name.over_react.g.dart';
@@ -13,14 +26,21 @@ part 'design_main_strand_domain_name.over_react.g.dart';
 UiFactory<DesignMainStrandDomainNameProps> DesignMainStrandDomainName = _$DesignMainStrandDomainName;
 
 mixin DesignMainStrandDomainNamePropsMixin on UiProps {
+  Strand strand;
   Domain domain;
   Helix helix;
   Geometry geometry;
+  BuiltMap<String, HelixGroup> helix_groups;
 
   int font_size;
   bool show_dna;
   String transform;
-  num svg_position_y;
+  Point<num> helix_svg_position;
+
+  List<ContextMenuItem> Function(Strand strand,
+      {@required Domain domain,
+      @required Address address,
+      @required ModificationType type}) context_menu_strand;
 }
 
 class DesignMainStrandDomainNameProps = UiProps with DesignMainStrandDomainNamePropsMixin;
@@ -29,9 +49,10 @@ class DesignMainStrandDomainNameComponent extends UiComponent2<DesignMainStrandD
     with PureComponent {
   @override
   render() {
-
-    Point<num> start_svg = props.helix.svg_base_pos(props.domain.start, props.domain.forward, props.svg_position_y);
-    Point<num> end_svg = props.helix.svg_base_pos(props.domain.end - 1, props.domain.forward, props.svg_position_y);
+    Point<num> start_svg =
+        props.helix.svg_base_pos(props.domain.start, props.domain.forward, props.helix_svg_position.y);
+    Point<num> end_svg =
+        props.helix.svg_base_pos(props.domain.end - 1, props.domain.forward, props.helix_svg_position.y);
     Point<num> mid_svg = (start_svg + end_svg) * 0.5;
 
     String baseline = props.domain.forward ? 'baseline' : 'hanging';
@@ -48,9 +69,41 @@ class DesignMainStrandDomainNameComponent extends UiComponent2<DesignMainStrandD
       ..x = '${mid_svg.x}'
       ..y = '${mid_svg.y}'
       ..dy = '${dy}'
+      ..id = id()
       ..transform = props.transform
       ..fontSize = props.font_size
       ..dominantBaseline = baseline
       ..className = constants.css_selector_domain_name_text)(props.domain.name);
+  }
+
+  String id() => props.domain.id + '_name';
+
+  // needed for capturing right-click events with React:
+  // https://medium.com/@ericclemmons/react-event-preventdefault-78c28c950e46
+  @override
+  componentDidMount() {
+    var element = querySelector('#${id()}');
+    element.addEventListener('contextmenu', on_context_menu);
+  }
+
+  @override
+  componentWillUnmount() {
+    var element = querySelector('#${id()}');
+    element.removeEventListener('contextmenu', on_context_menu);
+    super.componentWillUnmount();
+  }
+
+  on_context_menu(Event ev) {
+    MouseEvent event = ev;
+    if (!event.shiftKey) {
+      event.preventDefault();
+      event.stopPropagation();
+      Address address = util.find_closest_address(event, [props.helix], props.helix_groups,
+          props.geometry, {props.helix.idx: props.helix_svg_position}.build());
+      app.dispatch(actions.ContextMenuShow(
+          context_menu: ContextMenu(
+              items: props.context_menu_strand(props.strand, domain: props.domain, address: address).build(),
+              position: event.page)));
+    }
   }
 }
